♻️ Refactor pour l'adapter app au nouveau Mock
