+++ components/DetailCourse.tsx
@@ -1,12 +1,11 @@
-import { useEffect, useState } from "react";
 import { View, StyleSheet } from "react-native";
 import { Text, TouchableRipple } from "react-native-paper";
-import { useNavigation } from "@react-navigation/native";
 import ChronoTopDepart from "./ChornoTopDepart";
 import DetailTrajet from "./DetailTrajet";
 import MenuBurger from "./MenuBurger";
 import CardNumeroLine from "./CardNumeroLine";
 import Quotas from "./Quotas";
+import TypeCourse from "./TypeCourse";
 
 import Course from "../models/Course";
 
@@ -14,61 +13,49 @@ type Props = {
   course: Course;
 };
 
-const dateOfTest = new Date();
-
 export default function DetailCourse(props: Props) {
   const { course } = props;
 
-  const navigation = useNavigation();
-
-
-  const [dateTimeHd, setDateTimeHd] = useState(new Date());
-  const [dateTimeHa, setDateTimeHa] = useState(new Date());
-
-  // pour le test en attende de modification du Mock;
-  const dateTimeDeparture = "2023-06-09T07:33:14.719Z";
-  const datetimeArrival = "2023-06-09T07:59:56.333Z";
-
-  useEffect(() => {
-    setDateTimeHd(new Date(dateTimeDeparture));
-    setDateTimeHa(new Date(datetimeArrival));
-  }, []);
-
   const handleGotoAction = () => {
+    // appuie sur le composant
     switch (course.mission) {
-      case "BSC HDF": 
-        navigation.navigate("SaisiBsc", {course: course});
-      default: 
+      case "BSC HDF":
+        console.log("Basculement vers la page saisi");
+        break;
+      case "HLP VS":
+        console.log("afficher le link waze ou cordonner GPS");
+        break;
+      default:
         console.log("pas action pour ce type de mission");
     }
-  }
+  };
+
+  const isMissionBSC = (): boolean => {
+    return course.mission === "BSC HDF";
+  };
 
   return (
-    <TouchableRipple
-      onPress={handleGotoAction}
-    >
+    <TouchableRipple onPress={handleGotoAction}>
       <View style={style.container}>
         <View style={style.lineUp}>
           <View style={style.detailTime}>
-            <ChronoTopDepart
-              currentDatetime={dateOfTest}
-              datetimeArrival={dateOfTest}
-              datetimeDepart={dateOfTest}
-            />
-            <DetailTrajet
-              departureTime={dateTimeHd}
-              arrivalTime={dateTimeHa}
-              departureCity={course.depart}
-              arrivalCity={course.arrivee}
-            />
+            {course.infoHoraireCourse && (
+              <ChronoTopDepart
+                currentDatetime={new Date()} //TODO: a refactoriser.
+                datetimeArrival={course.infoHoraireCourse.datetimeArriveEnq}
+                datetimeDepart={course.infoHoraireCourse.datetimeDepartEnq}
+              />
+            )}
+            {course.infoHoraireCourse && (
+              <DetailTrajet infoHoraireCourse={course.infoHoraireCourse} />
+            )}
           </View>
-          <View>
-            <MenuBurger />
-          </View>
+          <View>{isMissionBSC() && <MenuBurger />}</View>
         </View>
         <View style={style.infoLine}>
+          <TypeCourse mission={course.mission} />
           {course.ligne && <CardNumeroLine lineNumber={course.ligne} />}
-          {course.prQuota && <Quotas value={course.prQuota} />}
+          {course.objectif && <Quotas value={course.objectif} />}
           <Text variant="headlineSmall">{course.trainCourse}</Text>
         </View>
       </View>
